add default template result processor
